+++ core/src/main/java/de/raysha/lib/jsimpleshell/completer/CommandNameCompleter.java
@@ -2,12 +2,16 @@ package de.raysha.lib.jsimpleshell.completer;
 
 import java.util.Collection;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 
 import jline.console.completer.Completer;
 import jline.console.completer.StringsCompleter;
 import de.raysha.lib.jsimpleshell.ShellCommand;
 import de.raysha.lib.jsimpleshell.Token;
+import de.raysha.lib.jsimpleshell.annotation.Param.DefaultTypes;
+import de.raysha.lib.jsimpleshell.completer.filter.CandidateFilter;
+import de.raysha.lib.jsimpleshell.completer.filter.CandidateFilter.FilterContext;
 import de.raysha.lib.jsimpleshell.handler.CommandAccessManager;
 import de.raysha.lib.jsimpleshell.handler.CommandAccessManager.AccessDecision;
 import de.raysha.lib.jsimpleshell.handler.CommandAccessManager.AccessDecision.Decision;
@@ -22,10 +26,15 @@ public class CommandNameCompleter implements Completer {
 
 	private CommandAccessManager accessManager;
 	private List<ShellCommand> commands;
+	private CandidateFilter filter;
 
 	public CommandNameCompleter(CommandAccessManager accessManager, List<ShellCommand> commands) {
 		this.accessManager = accessManager;
 		this.commands = commands;
+	}
+
+	public void setFilter(CandidateFilter filter) {
+		this.filter = filter;
 	}
 
 	@Override
@@ -48,7 +57,26 @@ public class CommandNameCompleter implements Completer {
 			}
 		}
 
+		filterCandidates(buffer, cursor, commandNames);
+
 		return new StringsCompleter(commandNames).complete(cutBuffer, cursor, candidates) + firstTokenIndex;
+	}
+
+	private void filterCandidates(String part, int cursor, Collection<String> commandNames) {
+		if(filter != null){
+			Iterator<String> iter = commandNames.iterator();
+			while(iter.hasNext()){
+				FilterContext ctx = createFilterContext(part, cursor, iter.next());
+
+				if(!filter.filter(ctx)){
+					iter.remove();
+				}
+			}
+		}
+	}
+
+	private FilterContext createFilterContext(String part, int cursor, String candidate) {
+		return new FilterContext(cursor, part, DefaultTypes.COMMAND_NAME, candidate);
 	}
 
 	private List<Token> getCurrentTokens(String buffer, int cursor) {