+++ integrationstests/src/test/java/de/raysha/lib/jsimpleshell/it/basic/AutoComplete.java
@@ -25,6 +25,7 @@ public class AutoComplete extends IntegrationsTest {
 						.addHandler(new CompleterCommands())
 						.addHandler(new ParamOrderCommands())
 						.addHandler(new SecurityCommands())
+						.enableAutocompleOfSpecialCommands()
 					.back();
 	}